+++ gcs/src/main/java/com/google/cloud/hadoop/fs/gcs/CoopLockFsckRunner.java
@@ -111,9 +111,10 @@ class CoopLockFsckRunner {
 
     Map<FileStatus, CoopLockRecord> expiredOperations =
         lockedOperations.stream()
-            .map(this::getOperationLockIfExpiredUnchecked)
+            .map(this::getOperationLockIfExpired)
             .filter(Optional::isPresent)
             .map(Optional::get)
+            .filter(this::checkLogExistsAndUnlockOtherwise)
             .collect(toMap(Map.Entry::getKey, Map.Entry::getValue));
 
     // If this is a check command then return after operations status was printed.
@@ -305,52 +306,35 @@ class CoopLockFsckRunner {
         .collect(toList());
   }
 
-  private Optional<Map.Entry<FileStatus, CoopLockRecord>> getOperationLockIfExpiredUnchecked(
+  private Optional<Map.Entry<FileStatus, CoopLockRecord>> getOperationLockIfExpired(
       CoopLockRecord operation) {
     try {
-      return getOperationLockIfExpired(bucketName, operation);
+      return getOperationLockIfExpiredChecked(operation);
     } catch (IOException e) {
       throw new RuntimeException(
           String.format("Failed to check if %s operation expired", operation), e);
     }
   }
 
-  private Optional<Map.Entry<FileStatus, CoopLockRecord>> getOperationLockIfExpired(
-      String bucketName, CoopLockRecord operationRecord) throws IOException {
+  private Optional<Map.Entry<FileStatus, CoopLockRecord>> getOperationLockIfExpiredChecked(
+      CoopLockRecord operationRecord) throws IOException {
     String globPath =
         CoopLockRecordsDao.LOCK_DIRECTORY + "*" + operationRecord.getOperationId() + "*.lock";
     URI globUri =
         gcsFs.getPathCodec().getPath(bucketName, globPath, /* allowEmptyObjectName= */ false);
-    FileStatus[] operationLocks = ghfs.globStatus(new Path(globUri));
+    FileStatus[] operationLock = ghfs.globStatus(new Path(globUri));
     checkState(
-        operationLocks.length < 2,
+        operationLock.length < 2,
         "operation %s should not have more than one lock file",
         operationRecord.getOperationId());
 
-    // Lock file not created - nothing to repair
-    if (operationLocks.length == 0) {
-      // Release lock if this is not a "check" command
-      // and if it processes this specific operation or "all" operations
-      if (!COMMAND_CHECK.equals(command)
-          && (ARGUMENT_ALL_OPERATIONS.equals(fsckOperationId)
-              || fsckOperationId.equals(operationRecord.getOperationId()))) {
-        logger.atInfo().log(
-            "Operation %s for %s resources doesn't have lock file, unlocking",
-            operationRecord.getOperationId(), operationRecord.getResources());
-        StorageResourceId[] lockedResources =
-            operationRecord.getResources().stream()
-                .map(resource -> new StorageResourceId(bucketName, resource))
-                .toArray(StorageResourceId[]::new);
-        lockRecordsDao.unlockPaths(operationRecord.getOperationId(), lockedResources);
-      } else {
-        logger.atInfo().log(
-            "Operation %s for %s resources doesn't have lock file, skipping",
-            operationRecord.getOperationId(), operationRecord.getResources());
-      }
+    // <operation>.lock file not created - nothing to repair
+    if (operationLock.length == 0) {
+      unlockOperationIfNecessary(operationRecord, "lock");
       return Optional.empty();
     }
 
-    FileStatus operationStatus = operationLocks[0];
+    FileStatus operationStatus = operationLock[0];
 
     if (operationRecord.getLockExpiration().isBefore(operationExpirationInstant)
         && getRenewedLockExpiration(operationStatus, operationRecord)
@@ -361,6 +345,64 @@ class CoopLockFsckRunner {
 
     logger.atInfo().log("Operation %s not expired.", operationStatus.getPath());
     return Optional.empty();
+  }
+
+  /**
+   * Returns {@code true} if operation log exists and {@code false} if it does not exist. As a
+   * sideffect releases operation lock if operation log does not exist.
+   */
+  private boolean checkLogExistsAndUnlockOtherwise(
+      Map.Entry<FileStatus, CoopLockRecord> operation) {
+    try {
+      return checkLogExistsAndUnlockOtherwiseChecked(operation);
+    } catch (IOException e) {
+      throw new RuntimeException(
+          String.format("Failed to check if %s operation expired", operation.getValue()), e);
+    }
+  }
+
+  private boolean checkLogExistsAndUnlockOtherwiseChecked(
+      Map.Entry<FileStatus, CoopLockRecord> operation) throws IOException {
+    CoopLockRecord operationRecord = operation.getValue();
+    String globPath =
+        CoopLockRecordsDao.LOCK_DIRECTORY + "*" + operationRecord.getOperationId() + "*.log";
+    URI globUri =
+        gcsFs.getPathCodec().getPath(bucketName, globPath, /* allowEmptyObjectName= */ false);
+    FileStatus[] operationLog = ghfs.globStatus(new Path(globUri));
+    checkState(
+        operationLog.length < 2,
+        "operation %s should not have more than one log file",
+        operationRecord.getOperationId());
+
+    // <operation>.log file not created - nothing to repair
+    if (operationLog.length == 0) {
+      unlockOperationIfNecessary(operationRecord, "log");
+      return false;
+    }
+
+    return true;
+  }
+
+  private void unlockOperationIfNecessary(CoopLockRecord operationRecord, String fileType)
+      throws IOException {
+    // Release lock if this is not a "check" command
+    // and if it processes this specific operation or "all" operations
+    if (!COMMAND_CHECK.equals(command)
+        && (ARGUMENT_ALL_OPERATIONS.equals(fsckOperationId)
+            || fsckOperationId.equals(operationRecord.getOperationId()))) {
+      logger.atInfo().log(
+          "Operation %s for %s resources doesn't have %s file, unlocking",
+          operationRecord.getOperationId(), operationRecord.getResources(), fileType);
+      StorageResourceId[] lockedResources =
+          operationRecord.getResources().stream()
+              .map(resource -> new StorageResourceId(bucketName, resource))
+              .toArray(StorageResourceId[]::new);
+      lockRecordsDao.unlockPaths(operationRecord.getOperationId(), lockedResources);
+    } else {
+      logger.atInfo().log(
+          "Operation %s for %s resources doesn't have %s file, skipping",
+          operationRecord.getOperationId(), operationRecord.getResources(), fileType);
+    }
   }
 
   private void deleteResource(String resource, Collection<String> loggedResources)