+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/CoopLockRepairIntegrationTest.java
@@ -70,7 +70,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Integration tests for {@link GoogleCloudStorageFileSystem} class. */
+/** Integration tests for Cooperative Locking FSCK tool. */
 @RunWith(JUnit4.class)
 public class CoopLockRepairIntegrationTest {
 
@@ -216,7 +216,7 @@ public class CoopLockRepairIntegrationTest {
 
     GoogleCloudStorageFileSystemOptions gcsFsOptions = newGcsFsOptions();
 
-    failDeleteOperation(bucketName, gcsFsOptions, dirUri);
+    failDeleteOperation(gcsFsOptions, bucketName, dirUri);
 
     GoogleCloudStorageFileSystem gcsFs = newGcsFs(gcsFsOptions, httpRequestInitializer);
 
@@ -261,18 +261,15 @@ public class CoopLockRepairIntegrationTest {
 
     GoogleCloudStorageFileSystemOptions gcsFsOptions = newGcsFsOptions();
 
-    failDeleteOperation(bucketName, gcsFsOptions, dirUri);
+    failDeleteOperation(gcsFsOptions, bucketName, dirUri);
 
     GoogleCloudStorageFileSystem gcsFs = newGcsFs(gcsFsOptions, httpRequestInitializer);
 
     // delete operation lock file
     List<URI> lockFile =
         gcsFs.listFileInfo(bucketUri.resolve(LOCK_DIRECTORY)).stream()
-            .filter(
-                f ->
-                    !f.getPath().toString().endsWith("/all.lock")
-                        && f.getPath().toString().endsWith(".lock"))
             .map(FileInfo::getPath)
+            .filter(p -> !p.toString().endsWith("/all.lock") && p.toString().endsWith(".lock"))
             .collect(toImmutableList());
     gcsFs.delete(Iterables.getOnlyElement(lockFile), /* recursive */ false);
 
@@ -302,6 +299,53 @@ public class CoopLockRepairIntegrationTest {
   }
 
   @Test
+  public void failedDirectoryDelete_noLogFile_checkSucceeds() throws Exception {
+    String bucketName = gcsfsIHelper.createUniqueBucket("coop-delete-check-no-log-failed");
+    URI bucketUri = new URI("gs://" + bucketName + "/");
+    String fileName = "file";
+    URI dirUri = bucketUri.resolve("delete_" + UUID.randomUUID() + "/");
+
+    // create file to delete
+    gcsfsIHelper.writeTextFile(bucketName, dirUri.resolve(fileName).getPath(), "file_content");
+
+    GoogleCloudStorageFileSystemOptions gcsFsOptions = newGcsFsOptions();
+
+    failDeleteOperation(gcsFsOptions, bucketName, dirUri);
+
+    GoogleCloudStorageFileSystem gcsFs = newGcsFs(gcsFsOptions, httpRequestInitializer);
+
+    // delete operation log file
+    List<URI> logFile =
+        gcsFs.listFileInfo(bucketUri.resolve(LOCK_DIRECTORY)).stream()
+            .map(FileInfo::getPath)
+            .filter(p -> p.toString().endsWith(".log"))
+            .collect(toImmutableList());
+    gcsFs.delete(Iterables.getOnlyElement(logFile), /* recursive */ false);
+
+    assertThat(gcsFs.exists(dirUri)).isTrue();
+    assertThat(gcsFs.exists(dirUri.resolve(fileName))).isTrue();
+
+    CoopLockFsck fsck = new CoopLockFsck();
+    fsck.setConf(getTestConfiguration());
+
+    fsck.run(new String[] {"--check", "gs://" + bucketName});
+
+    assertThat(gcsFs.exists(dirUri)).isTrue();
+    assertThat(gcsFs.exists(dirUri.resolve(fileName))).isTrue();
+
+    // Validate lock files
+    List<URI> lockFiles =
+        gcsFs.listFileInfo(bucketUri.resolve(LOCK_DIRECTORY)).stream()
+            .map(FileInfo::getPath)
+            .collect(toList());
+
+    assertThat(lockFiles).hasSize(2);
+    assertThat(matchFile(lockFiles, "all\\.lock")).isNotNull();
+    String filenamePattern = String.format(OPERATION_FILENAME_PATTERN_FORMAT, DELETE);
+    assertThat(matchFile(lockFiles, filenamePattern + "\\.log").isPresent()).isFalse();
+  }
+
+  @Test
   public void failedDirectoryDelete_rollForward_withWrongId_fails() throws Exception {
     String bucketName = gcsfsIHelper.createUniqueBucket("coop-delete-fwd-fail-bad-id");
     URI bucketUri = new URI("gs://" + bucketName + "/");
@@ -313,7 +357,7 @@ public class CoopLockRepairIntegrationTest {
 
     GoogleCloudStorageFileSystemOptions gcsFsOptions = newGcsFsOptions();
 
-    failDeleteOperation(bucketName, gcsFsOptions, dirUri);
+    failDeleteOperation(gcsFsOptions, bucketName, dirUri);
 
     GoogleCloudStorageFileSystem gcsFs = newGcsFs(gcsFsOptions, httpRequestInitializer);
 
@@ -363,7 +407,7 @@ public class CoopLockRepairIntegrationTest {
 
     GoogleCloudStorageFileSystemOptions gcsFsOptions = newGcsFsOptions();
 
-    failDeleteOperation(bucketName, gcsFsOptions, dirUri);
+    failDeleteOperation(gcsFsOptions, bucketName, dirUri);
 
     GoogleCloudStorageFileSystem gcsFs = newGcsFs(gcsFsOptions, httpRequestInitializer);
 
@@ -449,6 +493,71 @@ public class CoopLockRepairIntegrationTest {
         .isEqualTo(new DeleteOperation().setLockExpiration(null).setResource(dirUri.toString()));
     assertThat(gcsfsIHelper.readTextFile(bucketName, logFileUri.getPath()))
         .isEqualTo(dirUri.resolve(fileName) + "\n" + dirUri + "\n");
+  }
+
+  @Test
+  public void failedDirectoryRename_noLogFile_successfullyRepaired() throws Exception {
+    String bucketName = gcsfsIHelper.createUniqueBucket("coop-rename-back-failed-copy-nolog");
+    URI bucketUri = new URI("gs://" + bucketName + "/");
+    String dirName = "rename_" + UUID.randomUUID();
+    String fileName = "file";
+    URI srcDirUri = bucketUri.resolve(dirName + "_src/");
+    URI dstDirUri = bucketUri.resolve(dirName + "_dst/");
+
+    // create file to rename
+    gcsfsIHelper.writeTextFile(bucketName, srcDirUri.resolve(fileName).getPath(), "file_content");
+
+    GoogleCloudStorageFileSystemOptions gcsFsOptions = newGcsFsOptions();
+
+    // fail rename operation during log file creation
+    failRenameOperation(
+        srcDirUri,
+        dstDirUri,
+        gcsFsOptions,
+        r ->
+            "POST".equals(r.getRequestMethod())
+                && r.getUrl().toString().contains(".log")
+                && !r.getUrl().toString().contains("all.log"));
+
+    GoogleCloudStorageFileSystem gcsFs = newGcsFs(gcsFsOptions, httpRequestInitializer);
+
+    assertThat(gcsFs.exists(srcDirUri)).isTrue();
+    assertThat(gcsFs.exists(srcDirUri.resolve(fileName))).isTrue();
+    assertThat(gcsFs.exists(dstDirUri)).isFalse();
+    assertThat(gcsFs.exists(dstDirUri.resolve(fileName))).isFalse();
+
+    CoopLockFsck fsck = new CoopLockFsck();
+    fsck.setConf(getTestConfiguration());
+
+    // Wait until lock will expire
+    sleepUninterruptibly(COOP_LOCK_TIMEOUT);
+
+    fsck.run(new String[] {"--rollBack", "gs://" + bucketName, "all"});
+
+    assertThat(gcsFs.exists(dstDirUri)).isFalse();
+    assertThat(gcsFs.exists(dstDirUri.resolve(fileName))).isFalse();
+    assertThat(gcsFs.exists(srcDirUri)).isTrue();
+    assertThat(gcsFs.exists(srcDirUri.resolve(fileName))).isTrue();
+
+    // Validate lock files
+    List<URI> lockFiles =
+        gcsFs.listFileInfo(bucketUri.resolve(LOCK_DIRECTORY)).stream()
+            .map(FileInfo::getPath)
+            .collect(toList());
+
+    assertThat(lockFiles).hasSize(1);
+    String filenameFormat = String.format(OPERATION_FILENAME_PATTERN_FORMAT, RENAME);
+    URI lockFileUri = matchFile(lockFiles, filenameFormat + "\\.lock").get();
+
+    String lockContent = gcsfsIHelper.readTextFile(bucketName, lockFileUri.getPath());
+    assertThat(GSON.fromJson(lockContent, RenameOperation.class).setLockExpiration(null))
+        .isEqualTo(
+            new RenameOperation()
+                .setLockExpiration(null)
+                .setSrcResource(srcDirUri.toString())
+                .setDstResource(dstDirUri.toString())
+                .setCopySucceeded(false));
+    assertThat(matchFile(lockFiles, filenameFormat + "\\.log").isPresent()).isFalse();
   }
 
   @Test
@@ -553,8 +662,7 @@ public class CoopLockRepairIntegrationTest {
     HttpRequestInitializer failingRequestInitializer = newFailingRequestInitializer(failPredicate);
     GoogleCloudStorageFileSystem failingGcsFs = newGcsFs(options, failingRequestInitializer);
 
-    IOException e =
-        assertThrows(IOException.class, () -> failingGcsFs.rename(srcDirUri, dstDirUri));
+    Exception e = assertThrows(Exception.class, () -> failingGcsFs.rename(srcDirUri, dstDirUri));
     assertThat(e).hasCauseThat().hasCauseThat().hasMessageThat().endsWith("Injected failure");
   }
 
@@ -581,7 +689,7 @@ public class CoopLockRepairIntegrationTest {
 
     GoogleCloudStorageFileSystemOptions gcsFsOptions = newGcsFsOptions();
 
-    failDeleteOperation(bucketName, gcsFsOptions, dirUri);
+    failDeleteOperation(gcsFsOptions, bucketName, dirUri);
 
     GoogleCloudStorageFileSystem gcsFs = newGcsFs(gcsFsOptions, httpRequestInitializer);
 
@@ -617,7 +725,7 @@ public class CoopLockRepairIntegrationTest {
   }
 
   private static void failDeleteOperation(
-      String bucketName, GoogleCloudStorageFileSystemOptions gcsFsOptions, URI dirUri)
+      GoogleCloudStorageFileSystemOptions gcsFsOptions, String bucketName, URI dirUri)
       throws Exception {
     HttpRequestInitializer failingRequestInitializer =
         newFailingRequestInitializer(