+++ gcs/CHANGES.md
@@ -1,5 +1,8 @@
 ### 2.1.0 - 2019-XX-XX
 
+1.  Cooperative Locking FSCK tool: fix recovery of operations that failed before
+    creating an operation log file.
+
 ### 2.0.0 - 2019-08-23
 
 1.  Remove Hadoop 1.x support.